rm duplicated tx_indexer code
