+++ internal/tx_indexer/indexer.go
@@ -8,14 +8,14 @@ import (
 
 	"github.com/jackc/pgx/v5"
 	"github.com/sirupsen/logrus"
-	"github.com/vultisig/verifier/tx_indexer/pkg/rpc"
-	"github.com/vultisig/verifier/tx_indexer/pkg/storage"
-	"github.com/vultisig/verifier/types"
 	"golang.org/x/sync/errgroup"
 
 	vstorage "github.com/vultisig/verifier/internal/storage"
+	"github.com/vultisig/verifier/plugin/tx_indexer"
 	"github.com/vultisig/verifier/plugin/tx_indexer/pkg/graceful"
-	"github.com/vultisig/verifier/tx_indexer"
+	"github.com/vultisig/verifier/plugin/tx_indexer/pkg/rpc"
+	"github.com/vultisig/verifier/plugin/tx_indexer/pkg/storage"
+	"github.com/vultisig/verifier/types"
 )
 
 type FeeIndexer struct {
