--- tx_indexer/pkg/conv/conv.go
@@ -1,13 +1,0 @@
-package conv
-
-func Ptr[T any](v T) *T {
-	return &v
-}
-
-func FromPtr[T any](v *T) T {
-	if v == nil {
-		var zero T
-		return zero
-	}
-	return *v
-}
